add and delete user for project

+++ src/controllers/project.controllers.ts
@@ -1,5 +1,7 @@
 import { Request, Response } from 'express'
 import { ParamsDictionary } from 'express-serve-static-core'
+import { ObjectId } from 'mongodb'
+import HTTP_STATUS from '~/constants/httpStatus'
 import { IResponseMessage } from '~/interfaces/reponses/response'
 import { CreateProjectReqBody } from '~/interfaces/requests/Project.requests'
 import Project from '~/models/schemas/Project.schemas'
@@ -11,6 +13,10 @@ class ProjectController {
     res: Response
   ): Promise<Response<IResponseMessage<typeof Project>>> {
     try {
+      const userRole = req.decoded_authorization?.role
+      if (userRole !== 'admin') {
+        return res.status(403).json({ success: false, code: HTTP_STATUS.FORBIDDEN, message: 'Forbidden Access denied' })
+      }
       const result = await projectService.createProject(req.body)
       return res.status(201).json(result)
     } catch (error) {
@@ -23,6 +29,10 @@ class ProjectController {
     res: Response
   ): Promise<Response<IResponseMessage<typeof Project>>> {
     try {
+      const userRole = req.decoded_authorization?.role
+      if (userRole !== 'admin') {
+        return res.status(403).json({ success: false, code: HTTP_STATUS.FORBIDDEN, message: 'Forbidden Access denied' })
+      }
       const projectId = req.params.projectId
       const result = await projectService.updateProjectById(projectId, req.body)
       return res.json(result)
@@ -33,6 +43,10 @@ class ProjectController {
   }
   async delete(req: Request, res: Response): Promise<Response<IResponseMessage<typeof Project>>> {
     try {
+      const userRole = req.decoded_authorization?.role
+      if (userRole !== 'admin') {
+        return res.status(403).json({ success: false, code: HTTP_STATUS.FORBIDDEN, message: 'Forbidden Access denied' })
+      }
       const projectId = req.params.projectId
       const result = await projectService.deleteProjectById(projectId)
       return res.json(result)
@@ -43,6 +57,10 @@ class ProjectController {
   }
   async getProject(req: Request, res: Response): Promise<Response<IResponseMessage<typeof Project>>> {
     try {
+      const userRole = req.decoded_authorization?.role
+      if (userRole !== 'admin') {
+        return res.status(403).json({ success: false, code: HTTP_STATUS.FORBIDDEN, message: 'Forbidden Access denied' })
+      }
       const projectId = req.params.projectId
       const result = await projectService.getProjectById(projectId)
       return res.json(result)
@@ -53,9 +71,41 @@ class ProjectController {
   }
   async getAllProject(req: Request, res: Response): Promise<Response<IResponseMessage<typeof Project>[]>> {
     try {
+      const userRole = req.decoded_authorization?.role
+      if (userRole !== 'admin') {
+        return res.status(403).json({ success: false, code: HTTP_STATUS.FORBIDDEN, message: 'Forbidden Access denied' })
+      }
       const page = parseInt(req.query.page as string)
       const pageSize = parseInt(req.query.limit as string)
       const result = await projectService.getAllProject(page, pageSize)
+      return res.json(result)
+    } catch (error) {
+      const err: Error = error as Error
+      throw new Error(err.message)
+    }
+  }
+  async addParticipant(req: Request, res: Response): Promise<Response<IResponseMessage<typeof Project>>> {
+    try {
+      const projectId = req.params.projectId
+      const participant = req.params.participantId
+      const result = await projectService.addParticipant(projectId, new ObjectId(participant))
+      if (result.success === false) {
+        return res.status(409).json(result)
+      }
+      return res.json(result)
+    } catch (error) {
+      const err: Error = error as Error
+      throw new Error(err.message)
+    }
+  }
+  async deleteParticipant(req: Request, res: Response): Promise<Response<IResponseMessage<typeof Project>>> {
+    try {
+      const projectId = req.params.projectId
+      const participant = req.params.participantId
+      const result = await projectService.deleteParticipant(projectId, new ObjectId(participant))
+      if (result.success === false) {
+        return res.status(404).json(result)
+      }
       return res.json(result)
     } catch (error) {
       const err: Error = error as Error
